Switching site to consume terra-navigation package
