+++ packages/terra-framework-site/src/app/tests/TestsMenu.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
 import MenuList from '../common/menu/MenuList';
-import RoutingDelegate from '../common/navigation/routing/RoutingDelegate';
+import RoutingDelegate from 'terra-navigation/lib/routing/RoutingDelegate';
 import { siteConfigPropType } from '../../componentConfig';
 
 const propTypes = {
