+++ packages/terra-navigation/src/NavigationLayout.jsx
@@ -5,12 +5,11 @@ import {
 } from 'react-router-dom';
 
 import AppDelegate from 'terra-app-delegate';
-import { navigationConfigPropType, configHasMatchingRoute } from './routing/RoutingConfigUtils';
 import Layout from 'terra-layout';
+import { navigationConfigPropType, configHasMatchingRoute } from './routing/RoutingConfigUtils';
 
 const propTypes = {
   app: AppDelegate.propType,
-  enableHoverMenu: PropTypes.bool,
   header: PropTypes.element,
   children: PropTypes.element,
   location: PropTypes.object,
@@ -81,7 +80,7 @@ class NavigationLayout extends React.Component {
   }
 
   render() {
-    const { header, children, menu, menuText, routeConfig, location, enableHoverMenu } = this.props;
+    const { header, children, menu, menuText, routeConfig, location } = this.props;
 
     let menuComponent;
     if (configHasMatchingRoute(location.pathname, routeConfig.menuRoutes, this.state.size)) {
@@ -93,7 +92,6 @@ class NavigationLayout extends React.Component {
         header={this.decorateElement(header)}
         menu={menuComponent}
         menuText={menuText}
-        enableHoverMenu={enableHoverMenu}
       >
         {this.decorateElement(children)}
       </Layout>
