+++ packages/terra-navigation/src/wrappers/ContentWrapper.jsx
@@ -7,22 +7,15 @@ import BaseWrapper from './_BaseWrapper';
 const propTypes = {
   app: AppDelegate.propType,
   routingManager: PropTypes.object,
-  size: PropTypes.string,
 };
 
-class ContentWrapper extends React.Component {
-  render() {
-    const { app, routingManager } = this.props;
-
-    return (
-      <BaseWrapper
-        app={app}
-        routingManager={routingManager}
-        routes={routingManager.routeConfig.contentRoutes}
-      />
-    );
-  }
-}
+const ContentWrapper = ({ app, routingManager }) => (
+  <BaseWrapper
+    app={app}
+    routingManager={routingManager}
+    routes={routingManager.routeConfig.contentRoutes}
+  />
+);
 
 ContentWrapper.propTypes = propTypes;
