+++ packages/terra-navigation/src/wrappers/MenuWrapper.jsx
@@ -7,22 +7,15 @@ import BaseWrapper from './_BaseWrapper';
 const propTypes = {
   app: AppDelegate.propType,
   routingManager: PropTypes.object,
-  size: PropTypes.string,
 };
 
-class MenuWrapper extends React.Component {
-  render() {
-    const { app, routingManager } = this.props;
-
-    return (
-      <BaseWrapper
-        app={app}
-        routingManager={routingManager}
-        routes={routingManager.routeConfig.menuRoutes}
-      />
-    );
-  }
-}
+const MenuWrapper = ({ app, routingManager }) => (
+  <BaseWrapper
+    app={app}
+    routingManager={routingManager}
+    routes={routingManager.routeConfig.menuRoutes}
+  />
+);
 
 MenuWrapper.propTypes = propTypes;
