+++ packages/terra-navigation/src/wrappers/_BaseWrapper.jsx
@@ -10,30 +10,23 @@ import RoutingStack from '../routing/RoutingStack';
 const propTypes = {
   app: AppDelegate.propType,
   routingManager: PropTypes.object,
-  size: PropTypes.string,
   routes: PropTypes.any,
 };
 
-class BaseWrapper extends React.Component {
-  render() {
-    const { app, routingManager, routes } = this.props;
-
-    return (
-      <div style={{ height: '100%' }}>
-        <RoutingStack
-          size={routingManager.size}
-          navEnabled
-          app={app}
-          routeConfig={routes}
-          location={routingManager.location}
-          routingManager={routingManager}
-        >
-          <Redirect to={routingManager.routeConfig.navigation.index} />
-        </RoutingStack>
-      </div>
-    );
-  }
-}
+const BaseWrapper = ({ app, routingManager, routes }) => (
+  <div style={{ height: '100%' }}>
+    <RoutingStack
+      size={routingManager.size}
+      navEnabled
+      app={app}
+      routeConfig={routes}
+      location={routingManager.location}
+      routingManager={routingManager}
+    >
+      <Redirect to={routingManager.routeConfig.navigation.index} />
+    </RoutingStack>
+  </div>
+);
 
 BaseWrapper.propTypes = propTypes;
